Create badgeUpdater class
