+++ packages/sanes-chrome-extension/src/extension/background/externalHandler/index.ts
@@ -23,21 +23,6 @@ function generateErrorResponse(id: number | null, message: string): JsonRpcError
   };
 }
 
-function updateExtensionBadge(): void {
-  const isExtensionContext = typeof chrome !== 'undefined'; // needed for tests
-  if (!isExtensionContext) {
-    return;
-  }
-
-  const queueSize = RequestHandler.requests().length;
-  const badgeText = queueSize === 0 ? '' : `${queueSize}`;
-  const iconPath = queueSize === 0 ? 'assets/icons/icon128.png' : 'assets/icons/request128.png';
-
-  chrome.browserAction.setIcon({ path: iconPath });
-  chrome.browserAction.setBadgeBackgroundColor({ color: [0, 0, 0, 0] });
-  chrome.browserAction.setBadgeText({ text: badgeText });
-}
-
 export async function handleExternalMessage(
   signingServer: SigningServer,
   externalRequest: any, // eslint-disable-line @typescript-eslint/no-explicit-any
